Heat Transfer

+++ src/thejavapowder/Variables.java
@@ -66,7 +66,7 @@ public class Variables {
     int RandomNum = 0;
 
     byte[][] Map = new byte[Width][Height];// The particle type map
-    short[][] HMap = new short[Width][Height];// The Heat type map
+    float[][] HMap = new float[Width][Height];// The Heat type map
     short[][] VMap = new short[Width][Height];// The Voltage type map
     byte[][] PMap = new byte[Width][Height];// The Particle Properties Map
     //Elements Initialization
